Commit: added autosize attribute to the cfspreadsheet tag
In ACF, by default, the value of this attribute is true. The columns in a sheet resize to accommodate the contents. To avoid resizing the columns, set it to false.
Auto sizing can be relatively slow on large sheets.
The value has been set to false in the extension, for the sake of backward compatibility. Once tested thoroughly, we can set it to false.